Studio's main content component
