+++ src/pages/Studio/index.js
@@ -2,6 +2,7 @@ import React from 'react'
 import Studio from '../../ducks/Studio'
 import withBoundaries from './withBoundaries'
 import { parseUrl } from '../../ducks/Studio/url'
+import CtaJoinPopup from '../../components/CtaJoinPopup/CtaJoinPopup'
 
 export default withBoundaries(
   ({ settings, options, metrics, events, ...props }) => {
@@ -11,6 +12,6 @@ export default withBoundaries(
     sharedState.metrics = sharedState.metrics || metrics
     sharedState.events = sharedState.events || events
 
-    return <Studio {...props} {...sharedState} />
+    return <Studio topSlot={<CtaJoinPopup />} {...props} {...sharedState} />
   }
 )
